Add a "deploy_to_staging" rule to be used by automatic builds
Uploads files to an S3 bucket for nextstrain.org and optionally sends a
Slack alert to a configured webhook.

The S3 bucket defaults to s3://nextstrain-staging, the live staging
source, but it can be overridden for testing using snakemake's --config
or --configfile options.

The optional Slack webhook can be provided using the same config
options.  By default no message is sent.