Update ncov_trees-background_ru.md
I think there is a typo in English version in line 80: "Based on this tree, we would conclude that A & B closely related to each other, and D & E are closely related to each other. A & B are more closely related to C than they are to D & C". The last sentence should be: "A & B are more closely related to C than they are to D & E".  @cassiawag  To issue #178