Annotate nodes with identical nucleotide haplotypes to USA/WA1/2020
Adds a generic script to create a node data JSON that annotates whether each
node in a given tree has identical sequences to a given node's. This script
supports nucleotide and amino acid sequences.

Also, adds a rule to the Snakefile to annotate which nodes have nucleotide
haplotypes identical to the first strain identified in Washington
State (USA/WA1/2020) and a corresponding auspice config entry denoting this
categorical annotation.