Align sequences in smaller partitions to speed up builds
Aligning all available sequences with a single mafft process uses too much
memory. This commit introduces a Python script and Snakemake data-dependent
execution logic to partition sequences into smaller sets, align each set
individually, and aggregate the alignments into a single file.

Note that this commit also explicitly sets the number of threads to use for
augur align using Snakemake's threads keyword [1]. With the splitting of
alignments into multiple partitions, it is now possible to parallelize
alignments with multiple jobs using Snakemake's `-j` flag. It is faster to run 2
alignment jobs in parallel with 2 threads than to run 1 job at a time with 4
threads (3 min vs. 6 min on my MacBook). It is also dangerous to run multiple
jobs in parallel with Snakemake and the 'auto' threads option for augur align
because each job will try to use all available threads. Snakemake's threads
keyword is smart enough to only use the minimum of cores requested by the user
or available on the system.

[1] https://snakemake.readthedocs.io/en/stable/snakefiles/rules.html#threads