Regions: include root sequences
The subsampling would fail when missing the root sequences Wuhan-Hu-1/2019 + Wuhan/WH01/2019. This makes sure they are included via --include and then makes sure they aren't duplicated by replacing the "cat" call with combine-and-dedup-fastas.py. I did it this way in expectation of having a Asia region so that these two Wuhan sequences are sometimes part of context and sometimes part of focus but always need to be present.