Annotate the make_priorities rule with expected max memory usage
Setting a memory annotation on the rule lets Snakemake's scheduler be
smarter about concurrent jobs not overrunning memory (and thus
triggering Linux's OOM killer), similar to how the scheduler takes into
account CPU cores.  For this annotation to be used, a corresponding
change to ncov-ingest is needed to set the overall memory available
using the --resources flag.

Currently this rule uses a maximum of 2-3GB of memory per wildcard
instance of it on my local computer.