Reduce inputs to augur refine with a conditional input function
Checks whether the current region is empty (the default build) or one of the
specific regions and only returns the full multiple sequence alignment as input
to augur refine in the default case. This should reduce the size of inputs to
augur refine for regional builds to just those sequences in the subsampled
alignment and reduce the amount of memory required for the refine step.