Add subsampling for global build
Use the focused subsampling rule with a custom number of strains for the global
build compared to the regional builds. Skip the context subsampling and
combining of focus and context sequences for the global build since we only need
the focused subsampling.

This commit uses a parameter function (`_get_exclude_argument_by_wildcards`) to
dynamically include or exclude a command line argument based on wildcard values.