Move Nextstrain config file into main config directory
Consolidates workflow configuration files into the same `config/` directory
instead of splitting them up across config and profiles directories. This change
puts all similar configuration files in one place that should be easier to
discover and remember and also eliminates potential confusion about which YAML
config file is which when there are multiple in a single profile directory.

This commit also renames the "regions_config" file to the "nextstrain_config"
file to emphasize that the configuration is specific to the Nextstrain team.