Log stdout and stderr both to the terminal and to a log file
Uses shell redirects to combine stderr with stdout streams and pipes these
through tee to enable logging to a file while retaining the output on the
terminal. This latter output can be useful for watching logs in real time while
the former output is useful for debugging after the fact.

Note: this implementation uses the longer redirect form instead of the more
modern "|&" equivalent because older versions of bash (still common on Macs that
users have chosen not to upgrade yet ;) do not support this abbreviated
expression.