Prevent workflow from trying to adjust global metadata
Previously, the `export_all_regions` rule tried to expand the adjusted metadata
path for all regions including the global build. The global build does not have
an adjusted metadata file, though; it uses the original metadata file instead.

This commit fixes the problem by introducing a more generic function to get the
correct metadata file conditional on the region and using this function as a
base for an input function (the original `_get_metadata_by_wildcards`) and in a
list comprehension in the `export_all_regions` rule inputs. These functions live
in a new Snakemake file for shared functions, `common.smk`, following the
guidelines recommended by the Snakemake authors [1].

[1] https://github.com/snakemake-workflows/dna-seq-gatk-variant-calling/blob/master/rules/common.smk