Consistently use the same Python executable running Snakemake
Our shell rules inconsistently invoked Python as "python3" and "python",
both of which can be problematic.  In some environments, notably Linux,
"python" is almost always Python 2 and "python3" near-universally
exists.  In other environments, Anaconda in Windows with Python 3.8 and
no Python 2 install, "python3" doesn't exist and "python" is Python 3.

By trying to use the same Python executable that's running our Snakemake
workflow, we can side step the issue of finding the right name for the
executable we want.

Resolves PR #274 with an alternative solution after discussion.