Increase subsampling
Another pass to increase subsampling. 
- `global` went from 4725 to 2635
- `north-america` went from 1508 background 3682 foreground to 721 background 2373 foreground
- `europe` went from 1081 background 3781 foreground to 654 background 2536 foreground
Swapped `global` subsampling to `country year month`. With this at `division year month` we were having to go down to like `4` to get low enough counts and then this was dropping a bunch of January samples from Hubei and resulting in countries with more divisions having more samples.