Explicitly enable custom rules and use this for Nextstrain exports
Nextstrain-specific rules were originally included in the main Snakefile by
default. These rules referenced Nextstrain-specific parameters which then also
had to be defined in the default parameters file. This commit explicitly
introduces the concept of a `custom_rules` config entry that takes a list of
paths to Snakemake files that should be included at the end of the main
Snakefile. Then we use this config infrastructure to only include the Nextstrain
export rules in the Nextstrain profile and delete the Nextstrain-specific
parameters now that they are no longer referenced by default.

Note that "legacy clades" are still included in the default parameters as they
are part of the main workflow. If we decide to remove these from the default
workflow later, we can delete this parameter then.

Resolves #451