Throw errors for improperly defined mutually exclusive/required args
Produces an informative error message when the user either:
 - defines both `sequences_per_group` and `subsample_max_sequences`
 - defines neither `sequences_per_group` nor `subsample_max_sequences`

The first error addresses the mutually exclusive nature of the two arguments and
alerts the user that they need to change a specific subsampling scheme to fix
the issue.

The second error addresses the required natures of the arguments and alerts the
user that they need to define one of the two subsampling limits.