Allow users to define empty keys in subsampling schemes
Checks whether subsampling values are explicitly undefined (equal to None) and
only attempts to process those values that are not None. If a None value is
passed, we convert it to an empty string, so it does not appear in the
subsampling shell command.

These changes allow users to explicitly state that they do not define a specific
subsampling key. This functionality is important because of the way Snakemake's
config data structures are deep merged. It is possible for the default values of
a subsampling scheme to conflict with a user's own custom subsampling scheme of
the same name. To avoid conflicts of mutually exclusive keys, users can override
the default keys with empty values.