Update the message for subsample rule to reflect different arguments
Clarifies that the subsample wildcard refers to a specific named subsampling
scheme and explicitly lists all of the parameters used by a given subsampling
rule. This approach errs on the side of verbosity to provide a better sense of
what each subsampling rule is actually doing than looking at the augur filter
command itself.