Validate settings for a specific subsampling set when it's requested
The `_get_subsampling_settings` function returns either a complete subsampling
scheme (consistently of one or more sets of subsampling rules) or a specific
subsampling set. This return value depends on the presence of a `subsample`
wildcard. We only want to apply validation checks to the subsampling sets when
they have been specifically requested. This commit indents the original
validation logic by one level such that it only gets called when the current
Snakemake job has a `subsample` wildcard.