Warn users about conflicting subsampling schemes
Prints an informative warning when users have defined a subsampling scheme that
has the same name as one of the default schemes and different settings. The
warning lists all conflicting schemes, explains why the conflict is a problem,
provides a solution to avoid the conflicts, and alerts users that a future
release of the workflow will treat these conflicts as errors.

The use of the sleep function here is a workaround for the fact that Snakemake's
output will quickly scroll past any warning we provide. By sleeping for a few
seconds, users at least have a chance to see that there was a warning and can
scroll back to the read the whole message.