Add documentation for how config loading order occurs
This documentation is for our own team when maintaining this code in the future
and trying to understand why it works. There is also a good chance that
Snakemake will change its implementation in a way that breaks our current
logic. These comments should at least help us track down the changes.