Add frequencies panel to Auspice config
We have been exporting tip frequencies for all regional ncov builds as _tip-frequencies.json. However, these haven't been shown due to the absence of "frequencies" in the list of "panels" in the Auspice config. We now have enough temporal variation that I think it's warranted to include a frequencies view.