Write subsampling log output to a file
When I added support for logging output from all primary rules to files
in a `logs/` directory, I forgot to add a `log` directive to the
subsampling rule. This oversight causes logging output to be written to
the screen only and prevents users from easily debugging issues with
subsampling [1]. This commit addresses this issue by adding a `log`
directive to the subsampling rule.

[1] https://discussion.nextstrain.org/t/diagnosing-error-filtering-issues/149/5