Allow parameters to define ambiguous date exclusion criteria
This allows both the main filtering step as well as individual subsampling schemes to define if ambiguous dates are to be excluded, and if so, what level of ambiguity is to be excluded.

By having this in the subsampling steps, we allow flexibility so that (e.g.) we can allow ambiguous sequences for a particular subset of (important) samples.