Download sequences and metadata separately
Splits the `download` rule into two separate rules to allow users to
only download metadata or sequences as needed. For the convenience of
users who commonly request the `download` rule, this commit keeps a
`download` rule that refers to the two new separate download rules.

This commit also updates the dependencies of other rules to refer to the
metadata filename from the config instead of a specific rule (to match
the new preferred style for Snakemake rules).