Allow config-defined multiple sequence & metadata files
A common use case is the desire to start from multiple sequence / metadata files. For instance, a build starting with the GISAID data and a set of new (non-public) sequences.

This extends the workflow to allow `config["metadata"]` and `config["sequences"]` to be an array. If so, then an additional rule (`combine_input_metadata` and `combine_input_sequences`, respectively) are run before analysis steps. Metadata is combined via `scripts/combine_metadata.py` which respects the order of metadata files supplied, and thus will overwrite conflicting (non-empty) fields, such that the last file supplied takes precedence.

The longer-term plan is for augur to natively allow multiple metadata/sequence inputs, which would make these steps unnecessary.