Reduce memory used by sequence partition script
Although we don't use this sequence partitioning script anymore, we
might in the future. Its original implementation was a proof-of-concept
that tried to read all sequences into memory to calculate how many
partitions were needed. This commit replaces that wasteful approach with
a simpler streaming approach that only reads a single record into memory
at a time and uses a counter to track when a new partition is needed.