Add config param to limit active builds
Allows users to limit active builds from the command line through a
new top-level configuration parameter `active_builds`. This parameter
should be a command-delimited list of build names representing a subset
of all builds that would be run otherwise.

This new config parameter enables users to test workflows with a subset
of builds without needing to explicitly specify all of the workflows
outputs for those builds (which can be numerous).