Provide an explicit max date for tip frequencies
Prevent tip frequency pivots from extending too far into the future when
no maximum date is provided by providing an explicit max date. Users can
define the `max_date` in the frequencies section of the
configuration (as with `min_date`) or leave this field empty. When the
`max_date` is not defined by the user, it defaults to today's date.

This commit copies the `numeric_date` utility from TreeTime into the
common functions for the workflow. The cost of duplicating this code is
worth the benefit of not requiring Augur or TreeTime to be installed in
the user's environment when they run Snakemake. This allows users to run
the workflow with `--use-conda` from an environment without anything but
Snakemake installed.