Calculate antigenic distance with escape scores
Use escape scores from Greaney et al. 2021 [1] to identify putative
epitope sites in the RBD both. This commit adds two different distance
maps including a simple epitope sites map for only those positions with
excess total escape values across any of 23 sera and a weighted epitope
map with the average proportion of escape provided at each site averaged
across all 23 sera.

This commit also adds rules to the main workflow to calculate distances
from the root of a given tree in the spike protein with the two new
distance maps. These distances are reported as "epitope mutations" and
"weighted epitope mutations", respectively.

Note that this is a proof-of-concept and not intended to be used for
scientific applications until we better understand what the
corresponding values mean (if anything).

[1] https://www.biorxiv.org/content/10.1101/2020.12.31.425021v1