Filter out samples prescreened for S dropout
Including "S dropout" samples biases prevalence of B.1.1.7 upwards. Labeling of samples in GISAID seems to be happening best in the US, but I think helpful to have this included in general.

This additional filtering is included in the Nextstrain profile rather than the default parameters because I imagine other groups will want to look specifically at trees of B.1.1.7.