Allow subsampling of different input files
Given different input files (see previous commit), we wish to be able to use specific subsampling rules for each one.

Instead of creating wildcarded subsampling rules and introducing new syntax to the builds YAML, here we add columns to the metadata file specifying which their input. This allows individual subsampling blocks to specify which input they should act on in a simple and scalable way.
The "unified" metadata file is created at the same stage as the multiple sequence files are combined.