Use sequence index for augur filter rules
Adds a rule to generate an initial sequence index with the new augur index
command and updates the workflow's augur filter rules to use this index.

Note that this implementation requires the download of the full sequence set
even if the user has requested to start the workflow from the filtered alignment
(since subsampling also needs the sequence index).

This also temporarily changes the conda environment to point to the augur branch
with the new index command. As a result, this change only works currently with
the `--use-conda` flag.