Use sequence index for augur filter rules
Adds a rule to generate an initial sequence index with the new augur
index command and updates the workflow's augur filter rules to use this
index. This approach creates a new index for each input sequence and
also a separate index for the unified alignment used for subsampling.

Note that this implementation can require the download of the full sequence
set even if the user has requested to start the workflow from the
filtered alignment (since subsampling also needs the sequence index).
Future commits will address storage of the sequence index in S3 for
specific input origins.