Define inputs as a YAML array
This shifts the config-defined inputs (often in the builds.yaml
file) to be a YAML array so that we can guarantee the order
of inputs. The order matters for combining metadata. The schema
has been updated to enforce this, along with the tutorial and
nextstrain build.

We specifically allow config["inputs"] to be an (empty) string to
allow the case of using a command-line override to remove it.