Add support for downloading sequence index from S3
Adds a rule to download a sequence index for a given origin wildcard, if
it has been defined in the inputs. Also, adds a commented line for a
future sequence index that we store in S3. If the index can't be
downloaded, the original sequences will be downloaded and reindexed.