Fix snakemake errors for pipelines without nextalign
The pipeline was broken for builds without nextalign (e.g. my_profile/example) due to a quirk or snakemake (make?). 

Issue: the outputs from `rule build_align` differ if we have nextalign or not - only in the former case does `rules.build_align.output.translations` exist.  This file is used as input by `rule aa_muts_explicit`, but we only ask for the output of this rule if we are running nextalign (as we know it only exists in this case). However snakemake still tries to build the entire (?) graph and thus won't work for profiles which don't use nextalign. Changing the input to a function fixes this!