Add nextalign to the default conda environment
nextalign is now available through Bioconda for Linux and OS X, so we
can include it here as a default. With this change, nextalign will now
be available through the Docker image and the conda environment, so we
can move toward making nextalign the default behavior.

As a side note, we'll need to update the global Nextstrain conda
environment [1] to include nextalign, too.

[1] https://github.com/nextstrain/conda