Stream combined sequences to disk instead of loading into memory
Instead of reading all sequences into memory to identify duplicates and
then write out to disk, stream sequences from the multiple input files
to disk and check for duplicates on the fly with a hash of the sequence
content indexed by strain name. Hashes require far less memory than the
actual sequences, so we are able to output a full list of duplicate
sequences to stderr instead of the first duplicate.

Note that this implementation does not treat duplicate strains as
errors. We may wish to return an error code, if the original behavior is
preferable.