Benchmark all rules
Adds benchmark directives to all rules in the workflow, to help us
better understand where the workflow is spending the most time and
memory. Since memory is not currently logged by the `--stats` output, we
need this additional level of metadata for the workflow.

Note that benchmarks do not report memory on OS X (at least on older
Macs), so this information will come solely from Linux-based systems
like AWS Batch runs, etc.