Specify memory requirements for most rules
Adds `mem_mb` resource requirements for high-memory rules based on
benchmark rules and input sizes. The most common source of moderate
memory usage is loading metadata in augur (or other scripts) as a
dictionary with augur's `read_metadata` function. The most common source
of high memory usage is loading multiple sequence alignments as in the
`refine` and `ancestral` rules.

Multipliers for memory usage based on input file sizes are conservative,
erring on the side of requesting more memory than less to avoid hard
memory limits that can kill jobs. For example, at the time the main
metadata file for Nextstrain builds uses 165 MB on disk and ~2200 MB in
memory (as a Python dictionary), so the multipler used for rules that
load metadata is 15.