Add the new `data_provenance` field to Nextstrain's auspice JSONs
Adds the `data_provenance` field to Nextstrain's auspice config JSONs
and updates the minimum Augur version to 11.3.0 that adds support for
this new field to the auspice schema. This new field is not consumed by
Auspice yet, but it will be used by our builds and others to explicitly
display the data sources that enabled these builds.