Add support for curated contextual inputs
Modifies the path-or-URL logic for multiple inputs so users can start
from a local "filtered" input file (e.g., one downloaded from GISAID)
that represents a set of contextual sequences to use for a new analysis.
Adds an example profile with a build that contextualizes the example
Australian samples with a curated global subsampled set.

Assuming that the curated inputs will be compressed when downloaded by
the user, this commit also enables reading compressed sequences when
combining and deduplicating multiple inputs.