Calculate change in frequency for nodes in tree
Adds a script, Snakemake rule, and configs to calculate the change in
frequency (i.e., "delta frequency") for nodes in a given tree over a
fixed period of time. The default time period is 4 weeks which
corresponds to 4 timepoints or "pivots" in the tip frequencies data. The
workflow exports this new `delta_frequency` annotation in auspice JSONs
for visualization as a color-by or scatterplot attribute.

Fixes #594