Exclude tips and zero-values delta frequencies
Do not annotate change in frequency for tips by default (but allow users
to override this with a flag) and do not emit entries for nodes with
zero-valued delta frequencies. The resulting new annotations only color
internal nodes and branches with non-zero delta frequency values which
allows us to more rapidly identify clades of interest.