Add logistic regression as an option for delta frequency calculations
Adds a `--method` argument to the delta frequency calculation script and
logic to calculate a regression from the logistic transform of
frequencies. The resulting annotation estimates logitic growth for
clades. This new annotation is included as "logistic_growth" in the
auspice JSON along with the original linear annotation, for comparison.