Treat empty focal alignment as an error
An empty focal alignment is most likely an error. In the best case,
using proximity-based subsampling is a waste of time when this happens.
In the worst case, the user has a misconfigured subsampling scheme that
they need to troubleshoot. Either way, we should let the user know what
is happening.

The worst case error is actually what happened when our CI tests
initially failed with this new code. The CI test's had an improperly
configured subsampling scheme such that the focal alignment was always
empty except for the two reference sequences. The additional check to
ignore reference sequences in this PR surfaced this bug in the CI.