Fix subsampling for CI tests
Fixes a bug in the CI profile's subsampling caused by its use of a fixed
set of older data and the Nextstrain profile's use of time-based
sampling. Now that our CI data is quite old, none of it passes the
time-based filters of the Nextstrain profile. This commit defines a
simpler subsampling scheme for the European build, so the CI test can
still check the proximity-based subsampling part of the workflow without
any time-specific filtering.

As a benefit of this new subsampling scheme, the CI tests complete in <1
minute instead of ~8-10 minutes.