Fix bug in CI when the root node does not have enough tips to calculate logistic growth
The latest CI builds have fewer samples than previous CI builds, so the
new requirement that nodes have >=50 tips to calculate logistic growth
is never true for the CI data. This leads to a bug in the logistic
growth script where the root node tries to inherit its logistic growth
value from a parent it doesn't have. This commit adds a specific check
for the existence of a parent node before trying to pull from that node
and assigns a default missing value in the edge case where the parent
does not exist.