Simplify and update the Conda environment
mafft 7.471 did not support streaming from stdin while 7.475 has no
issues with this. While we are updating mafft, we can also require the
latest nextalign and Augur versions.

Uses the Augur Bioconda package allowing us to remove most dependencies
that are handled by that package, explicitly pins the latest version of
IQ-TREE 2 so users can benefit from the new version even if they already
have IQ-TREE installed, and removes packages like rethinkdb, nodejs,
pip, etc. that are not necessary in the context of a Snakemake run.