Report Augur versions on error without use_conda
If the workflow errors out and the user isn't using the `--use-conda`
Snakemake flag, let the user know what Augur version they have installed
and which version is expected in the Conda environment. Let users know
about the `--use-conda` flag in this case as a possible solution to the
workflow errors.