Define default build with no subsampling
Replaces the default "global" build when users don't provide any build
information with a "default-build" that uses a new built-in subsampling
scheme of "all" that does not perform any subsampling.

This change allows the default build to do "the right thing" for most
users which is to build a tree from all of the data in the inputs
they've defined from GISAID, INSDC, etc. curation.