Assume no subsampling when a scheme is not defined
When a build has no subsampling scheme defined, uses the new "all"
scheme to effectively skip subsampling. Prints a warning message to the
user, in case they don't realize they haven't defined a scheme.

This could be better. Perhaps it would be useful to tell the user that
they can explicitly select the "all" subsampling scheme?