Infer compression for downloaded metadata
Fixes a bug in the download_metadata rule where the rule hardcoded
`gunzip` as the decompression tool but gunzip is not guaranteed to
support decompression of LZMA-compressed files (`.xz` files) on all
operating systems (or at least across all implementations of `gunzip`).
This commit uses the same pattern as the other download rules of
inferring the decompression command from the filename.