Document all available configuration parameters
Documents all available configuration parameters in a single file,
following the general format of JSON Schema [1] so we can eventually
encode this information in the schema and autogenerate docs from that
file, if we choose.

Adds links to this new documentation from both the README and the
tutorial index file. As part of these edits, also cleans up the README
and index file tables of contents.

Relates to #632

[1] https://json-schema.org/