Fix bug in configuration docs
The min and max date filters for subsampling require the complete augur
filter argument instead of just the date to filter by. This is an
inconsistency in the subsampling logic with the filter logic which is
really a bug in the workflow, but we should at least document what
works.