Compress all files before S3 upload
Updates rules to use xz compression instead of gz (for consistency with
current S3 uploads), removes compression logic from the S3 upload script
and rule, and updates all outputs that are uploaded to S3 to be
compressed prior to upload. This change addresses the potential
complexity of an input to the upload rule being compressed in one format
and being uploaded in another, etc. by just compressing everything prior
to upload.

This commit updates the mutation summary and diagnostic scripts to
support compressed inputs and outputs.

Note that the mutation summary script hardcodes filename suffixes with
compression details that might easily change in the workflow and be
missed by developers. We should make this script more robust to changing
inputs to avoid silent failures in the future.