Calculate Jaccard distances between strains
Uses Jaccard distance between sets of amino acid substitutions annotates
in GISAID metadata as an alternative to the SNP-based proximity
calculation. This approach does not require any alignment or interaction
with sequence data, since the relevant information is already provided
in the metadata.

Adds a top-level parameter of which field to look for in the metadata
for comma-delimited substitutions to compare with Jaccard distances.
This parameter might make more sense in the subsampling definitions,
though.