Generate auspice configs via a rule
Internal nextstrain workflows typically generate many datasets. Currently we tend to use a single auspice config JSON for each dataset, despite these configs being essentially identical. Furthermore, the "nextstrain-open" profile was using the config files from the main (GISAID) profile, which were not well suited to the metadata available for the open builds. (Note that the config file deleted in this commit was never being used.)

Here we move to generating the auspice configs via a rule, which has a number of advantages. It is now easy to get an overview of the config fields which are the same, and which ones are different across builds in a profile; comments (which are allowed since it's javascript) also help with understanding. A rule allows us to easily have different settings for different builds, and generation may become dynamic in the future. Finally it helps prevent different build config files diverging unintentionally.

Currently this is implemented for the nextstrain-open profile, but future work will extend this to the (GISAID) nextstrain profile. There will be a common "base" config which can be imported by both in this case.

For users running few builds, it's preferable to avoid this complexity and stick with the config-files approach we currently describe in the tutorials. Workflows with many targets may wish to add their own rules similar to that done here.