Launch both GISAID and Open builds for trials
Separate workflow jobs so that they can be independently managed in the
GitHub Actions UI.

Copy and pasted "nextstrain build" invocations (instead of, e.g., a
shared script or YAML anchor) so they can be independently tweaked as
needed in the future.  For example, I'm starting with the same resources
for each, but that's probably unnecessary right now and we may want to
tune it sooner than later.