Update GISAID-based data prep instructions
Adds two GISAID-specific approaches to preparing data for a common use
case of running a "regional build" (e.g., a Washington State build).
These approaches include 1) searching for strains through GISAID and
downloading prebuilt contextual sequences from the GISAID "nextregions"
downloads and 2) downloading the full GISAID sequences and metadata and
filtering these locally to a subset of desired strains. These new
recommended approaches are possible as of v7 of the ncov workflow and
are based on feedback from partners in public health labs.

As part of restructuring the data prep page for clarity, this commit
also adds a new "Reference guides" section of docs, moves the
existing workflow configuration parameters guide to this new section,
and moves content about metadata fields from an appendix of the data
prep page to its own reference page where it is easier to discover and
link to.