Use subsampled metadata for per-build outputs
Replaces calls to unified metadata (or metadata-by-wildcards) with an
explicit path to the adjusted subsampled metadata from the final
subsampling filter step. The subsampled metadata should be substantially
smaller than the full unified metadata, so the rules modified in this
commit should all run correspindingly faster.

The use of the full metadata was a holdover from earlier versions of the
workflow that did not produce a subsampled metadata. The use of the full
metadata in the rules modified here was an oversight of mine when I
added the subsampled metadata output.

Fixes #654