Fix bug when adjusting metadata for the global build
This commit passes a default region of "global" when no region is defined in a
build and fixes a bug in the Nextstrain "global" build that does not define a
"region" attribute.

By using the adjusted metadata for most downstream processes, we bypass
conditional Snakemake logic that used to skip metadata adjustment for builds
named "global" or builds without any "region" attribute. Since the adjust
metadata script doesn't change the metadata when the given region isn't in the
list of main continental regions, we can skip the conditional Snakemake logic
and let this logic be handled internally to the adjust metadata script. In this
way, we always produce an "adjusted metadata" file, but it is not always
modified from its original input.